fix(deals): the access verbs reach every gate, and two files split at their seam
Gate findings from the participants slice.

`invite` and `revoke` now render an authorization rule (deal_room.update, the
grant their write paths take) and a history phrase — "invited" and "revoked
access for" rather than created/archived, because the participant row is
incidental and the access is the fact. Without the first, every one of those
audit rows would have recorded a blank governing grant.

Two files crossed the 500-line cap and are split where their concepts already
divided, not waived. serveroptions.go gives up the schema-authority group — the
owner-DSN pool and the non-production reset — which is the one group where
mis-wiring is destructive rather than merely absent. catalog.go gives up the
consumer groups: which types exist is a different question from who reads them,
and the two change for different reasons.

lifecycle_test.go stopped reading crm.yaml with yaml.v3, which the module is not
granted and arch-lint refused. It parses the generated DealRoomState constants
instead — and that matters beyond the lint: the intermediate version I wrote used
a hand-written candidate list, which silently could NOT catch an added state.
Mutation-checked both ways; the version here fails on a tenth state, the one it
replaced passed.
